All smoke tests for all commands now pass
In short: ReportingCommand.phase is back because Jacob's not yet implemented the streaming=True metadata item that's intended to specify whethere we are in the map or the reduce phase.

Debugging logic is also improved. See changes to app.py.

Next Up: Ensure that ReportingCommand. configuration is set based on ReportingCommand.phase.