Recorder now implements readline method and tempfile.tempdir is now correctly set
*Discovery: *
The Python 2.7 library does not expose the primitives that the C code uses. For example, file.readline does not call file.read. We must implement all methods that are required to record inputs and outputs. We now implement all that we must implement to support the External Search Command protocol.