Recording work with some caveats noted in search_command.py
In brief, recordings are currently useful if and only if you don't need access to the dispatch directory or those bits of SearchCommand.metadata that are ephemeral or a function of the environment under test. Recordings would be more useful if we had a set of mocks to these SearchCommand properties:
    SearchCommand.metadata,
    SearchCommand.search_results_info, and
    SearchCommand.service.

We save three timestamped files: an input file, an output file, and a snapshot of the dispatch_dir because it provides a good baseline for mocks that we or customers might develop.

Filename format:
<class-name>-<timestamp>.[input|output|dispatch_dir.tar.gz]